requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask_MySQLdb==2.0.0
matplotlib==3.8.2
PyMySQL==1.1.1
numpy==1.26.4
opencv_python==4.8.1.78
